#244 check the JQ version before converting the instances

+++ script/instances-0.6.0-to-0.7.0.sh
@@ -7,25 +7,40 @@ fi
 
 set -euo pipefail
 
-instanceDir="$1"
-backupDir="$instanceDir.bak_$(date +%s)"
-
-echo "Backing up $instanceDir to $backupDir."
-cp -r $instanceDir $backupDir
-echo "Converting instances format in $instanceDir from Broccoli <0.6.0 to 0.7.0."
-for instanceFile in $instanceDir/*.json; do
-  echo "- Converting $instanceFile"
-  tmpInstanceFile="$instanceFile.tmp"
-  cat $instanceFile | jq '.template.parameterInfos = (.template.parameterInfos | with_entries(.value.id = .value.name))' > $tmpInstanceFile
-  mv $tmpInstanceFile $instanceFile
-done
-
-echo "Conversion finished. Looks like everything went well."
-read -p "Delete $backupDir? [y/n] " -n 1 -r
-echo
-if [[ $REPLY =~ ^[Yy]$ ]]; then
-  echo "Deleting $backupDir."
-  rm -rf $backupDir
+jq_version="$(jq --version 2>&1)" # redirecting stderr https://github.com/stedolan/jq/issues/1452
+continue=0
+if [[ $jq_version != *"1.5"* ]]; then
+  echo "WARNING: This script was tested against jq-1.5 and should be run against that as well."
+  read -p "WARNING: Looks like you are using $jq_version. Continue anyway? [y/n] " -n 1 -r
+  echo
+  if [[ $REPLY =~ ^[Yy]$ ]]; then
+    continue=1
+  fi
 else
-  echo "Keeping $backupDir for now. You have to delete it manually."
+  continue=1
+fi
+
+if [[ "$continue" == 1 ]]; then
+  instanceDir="$1"
+  backupDir="$instanceDir.bak_$(date +%s)"
+
+  echo "Backing up $instanceDir to $backupDir."
+  cp -r $instanceDir $backupDir
+  echo "Converting instances format in $instanceDir from Broccoli <0.6.0 to 0.7.0."
+  for instanceFile in $instanceDir/*.json; do
+    echo "- Converting $instanceFile"
+    tmpInstanceFile="$instanceFile.tmp"
+    cat $instanceFile | jq '.template.parameterInfos = (.template.parameterInfos | with_entries(.value.id = .value.name))' > $tmpInstanceFile
+    mv $tmpInstanceFile $instanceFile
+  done
+
+  echo "Conversion finished. Looks like everything went well."
+  read -p "Delete $backupDir? [y/n] " -n 1 -r
+  echo
+  if [[ $REPLY =~ ^[Yy]$ ]]; then
+    echo "Deleting $backupDir."
+    rm -rf $backupDir
+  else
+    echo "Keeping $backupDir for now. You have to delete it manually."
+  fi
 fi
